backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
google-genai==1.0.0
pydantic-settings==2.4.0
python-dotenv==1.0.0